+++ apps/api/src/app/portfolio/calculator/roai/portfolio-calculator-novn-buy-and-sell-partially.spec.ts
@@ -129,15 +129,22 @@ describe('PortfolioCalculator', () => {
         groupBy: 'month'
       });
 
+      const investmentsByYear = portfolioCalculator.getInvestmentsByGroup({
+        data: portfolioSnapshot.historicalData,
+        groupBy: 'year'
+      });
+
       expect(portfolioSnapshot).toMatchObject({
         currentValueInBaseCurrency: new Big('87.8'),
         errors: [],
         hasErrors: false,
         positions: [
           {
+            activitiesCount: 2,
             averagePrice: new Big('75.80'),
             currency: 'CHF',
             dataSource: 'YAHOO',
+            dateOfFirstActivity: '2022-03-07',
             dividend: new Big('0'),
             dividendInBaseCurrency: new Big('0'),
             fee: new Big('4.25'),
@@ -197,6 +204,10 @@ describe('PortfolioCalculator', () => {
       expect(investmentsByMonth).toEqual([
         { date: '2022-03-01', investment: 151.6 },
         { date: '2022-04-01', investment: -75.8 }
+      ]);
+
+      expect(investmentsByYear).toEqual([
+        { date: '2022-01-01', investment: 75.8 }
       ]);
     });
   });