+++ apps/api/src/app/portfolio/calculator/roai/portfolio-calculator-novn-buy-and-sell.spec.ts
@@ -129,6 +129,11 @@ describe('PortfolioCalculator', () => {
         groupBy: 'month'
       });
 
+      const investmentsByYear = portfolioCalculator.getInvestmentsByGroup({
+        data: portfolioSnapshot.historicalData,
+        groupBy: 'year'
+      });
+
       expect(portfolioSnapshot.historicalData[0]).toEqual({
         date: '2022-03-06',
         investmentValueWithCurrencyEffect: 0,
@@ -188,9 +193,11 @@ describe('PortfolioCalculator', () => {
         hasErrors: false,
         positions: [
           {
+            activitiesCount: 2,
             averagePrice: new Big('0'),
             currency: 'CHF',
             dataSource: 'YAHOO',
+            dateOfFirstActivity: '2022-03-07',
             dividend: new Big('0'),
             dividendInBaseCurrency: new Big('0'),
             fee: new Big('0'),
@@ -248,6 +255,10 @@ describe('PortfolioCalculator', () => {
       expect(investmentsByMonth).toEqual([
         { date: '2022-03-01', investment: 151.6 },
         { date: '2022-04-01', investment: -151.6 }
+      ]);
+
+      expect(investmentsByYear).toEqual([
+        { date: '2022-01-01', investment: 0 }
       ]);
     });
   });