+++ libs/common/src/lib/interfaces/responses/portfolio-holding-response.interface.ts
@@ -11,11 +11,11 @@ export interface PortfolioHoldingResponse {
   activitiesCount: number;
   averagePrice: number;
   dataProviderInfo: DataProviderInfo;
+  dateOfFirstActivity: string;
   dividendInBaseCurrency: number;
   dividendYieldPercent: number;
   dividendYieldPercentWithCurrencyEffect: number;
   feeInBaseCurrency: number;
-  firstBuyDate: string;
   grossPerformance: number;
   grossPerformancePercent: number;
   grossPerformancePercentWithCurrencyEffect: number;