+++ libs/ui/src/lib/environment/environment.interface.ts
@@ -1,5 +1,4 @@
 export interface GfEnvironment {
   lastPublish: string | null;
   production: boolean;
-  stripePublicKey: string;
 }